US2 : create ad details (#3)
* US2 : update navigation with app router

* US2: create ad details page and image pager.

* US2: Update Tus

---------

+++ lib/domain/entities/ad/post_ad_entity.dart
@@ -10,7 +10,7 @@ class PostAdEntity with _$PostAdEntity {
   const factory PostAdEntity(
       {required final String title,
       required final String description,
-      final int? price,
+      final double? price,
       required final AdType adType,
       required final List<PhotoEntity> photos,
       required final CityEntity city,
